requirements.txt
langchain==0.1.14
langchain-openai==0.1.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.18.1
faiss-cpu==1.8.0
altair==4.0.0
tiktoken==0.6.0


#-e .
